molcom/requirements.txt
Django==1.6.1
South==0.8.4
django-filter==0.7
djangorestframework==2.3.10
phonenumbers==5.9.1
wsgiref==0.1.2
